// src/common/enum/permission.enum.ts
// perm for admin 
export enum PermissionType {
  MANAGE_USERS = 'manage_users',
  MANAGE_PRODUCTS = 'manage_products',
  MANAGE_ORDERS = 'manage_orders',
  MANAGE_CATEGORIES = 'manage_categories',
  MANAGE_DISCOUNTS = 'manage_discounts',
  MANAGE_REVIEWS = 'manage_reviews',
  MANAGE_SETTINGS = 'manage_settings',
  VIEW_ANALYTICS = 'view_analytics',
  MANAGE_CONTENT = 'manage_content',
  MANAGE_PROMOTIONS = 'manage_promotions',
  MANAGE_INVENTORY = 'manage_inventory',
  MANAGE_SHIPPING = 'manage_shipping',
  MANAGE_PAYMENTS = 'manage_payments',
  MANAGE_CUSTOMERS = 'manage_customers',
  MANAGE_REPORTS = 'manage_reports',
  MANAGE_SUPPORT = 'manage_support',
  MANAGE_AFFILIATES = 'manage_affiliates',
  MANAGE_LOYALTY_PROGRAMS = 'manage_loyalty_programs',
  MANAGE_SUBSCRIPTIONS = 'manage_subscriptions',
  MANAGE_BLOG = 'manage_blog',
}